feat(ci): Include backend-parity tests in host coverage run
`./ci.sh --coverage` now builds note-cpp-integration-backends with --coverage flags and merges its tracefile into the unified lcov. cjson.hpp and nlohmann.hpp were previously absent from host coverage because no unit test exercised their paths.

Coverage deltas:
- Previously:  lines 97.5% / functions 99.0% / branches 96.2%
- Now:         lines 96.4% / functions 98.2% / branches 94.4%
(aggregate drops slightly because the backend headers add real code with partial coverage — the headers are now IN the report, not filtered out.)

Host coverage for backend headers:
- cjson.hpp:    78/140 lines (was 0 — file was skipped)
- nlohmann.hpp: 64/119 lines (was 0 — file was skipped)

All thresholds still met (lines >= 90, functions >= 90, branches >= 94).

Drive-by fix: cjson.hpp had two long-long -> double implicit conversions that g++-13 with -Wconversion -Werror rejected. Added static_cast<double> in both spots. Clang accepts the implicit conversion so this hadn't surfaced.

Part of Phase 1: docs/superpowers/plans/2026-04-24-backend-parity-pilot.md

+++ src/note/backends/cjson.hpp
@@ -70,7 +70,7 @@ public:
         return *this;
     }
     CjsonBuilder& add(string_view key, json_int_t value) override {
-        cJSON_AddItemToObject(current(), zkey(key), cJSON_CreateNumber(value));
+        cJSON_AddItemToObject(current(), zkey(key), cJSON_CreateNumber(static_cast<double>(value)));
         return *this;
     }
     CjsonBuilder& add(string_view key, double value) override {
@@ -119,7 +119,7 @@ public:
         return *this;
     }
     CjsonBuilder& add_element(json_int_t value) override {
-        cJSON_AddItemToArray(current(), cJSON_CreateNumber(value));
+        cJSON_AddItemToArray(current(), cJSON_CreateNumber(static_cast<double>(value)));
         return *this;
     }
     CjsonBuilder& add_element(double value) override {
